requirements.txt
channels==2.1.6
channels-redis==2.3.1
Django==2.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-gis==0.14
django-cors-headers==2.4.0
django-filter==2.0.0
dj-database-url==0.5.0
factory-boy==2.11.1
flake8==3.7.7
Faker==1.0.7
paho-mqtt==1.4.0
Pillow==5.3.0
psycopg2-binary==2.7.6.1
pytest-cov==2.6.0
pytest==4.0.1
pytest-django==3.4.4
